apis: Fix UnixTimestampMilliseconds.to_representation
The implementation was wrong. `timedelta.total_seconds()` has
microsecond accuracy. Multiplied by 1000, we had a number of
milliseconds. But the implementation added `td.microseconds / 1000`,
which was wrong. A similarly correct implementation was:

    td = value - datetime.datetime(1970, 1, 1, tzinfo=datetime.timezone.utc)
    return round(td.total_seconds() * 1000)

... but `round(value.timestamp() * 1000)` is shorter.

I took the opportunity to simplify the inverse function (conversion
from an MDS timestamp to a datetime object). The previous
implementation was correct but there is a simpler way with
`datetime.fromtimestamp()`.