poll_providers: persist token in Django cache (#93)
Indeed we were caching the token for the life of the polling session,
but as a cron job, a new session was started every minute or so...

The token is considered as a password and is encrypted. Not with
Django's secret key, the required key is a particular format.

Host projects are free to choose the cache of their choice.
And obviously change how the encryption token is read.