models: fix EventRecord default saved_at timestamp type
We used the fancy https://docs.djangoproject.com/fr/2.2/ref/contrib/postgres/functions/#transactionnow
it returns the START of the database transaction.
This is actually not a desirable behaviour for a `saved_at` field. We want the time
at which it was recorded, not when the enclosing transaction begun, which may be
an hour ago for example with a very long job.

Note: I think TransactionNow would be useful if we had a timestamp to
assess the time validity of an object. In that case, knowing when the transaction
that wrote it started would indeed be nice.